refactor: move authentication logic to /modules/commands
