+++ src/app/api/auth/register/route.ts
@@ -1,43 +1,27 @@
 import { NextRequest, NextResponse } from 'next/server';
-import { hash } from 'bcryptjs';
-import { UserSchema, PAPSchema } from '@/modules/business-types';
+import { UserSchema } from '@/modules/business-types';
 import { getDb } from '@/modules/mongodb';
+import { handler$Register } from '@/modules/commands/Authenticate/handler';
 
 export async function POST(req: NextRequest) {
     try {
         const body = await req.json();
-        const { firstName, lastName, email, password } = UserSchema.parse(body);
+        const parsedBody = UserSchema.safeParse(body);
 
-        const db = await getDb();
-
-        const existingUser = await db.collection('users').findOne({ email });
-
-        if (existingUser) {
+        if (!parsedBody.success) {
             return NextResponse.json(
-                { message: 'User already exists' },
+                {
+                    error: parsedBody.error.message || "invalid params",
+                },
                 { status: 400 }
             );
         }
 
-        const hashedPassword = await hash(password, 10);
+        const db = await getDb();
 
-        const newUser = await db.collection('users').insertOne({
-            firstName,
-            lastName,
-            email,
-            password: hashedPassword,
-            role: 'user',
-        });
+        const response = await handler$Register(db, parsedBody.data);
 
-        await db.collection('paps').insertOne(
-            PAPSchema.parse({ userId: newUser.insertedId })
-        );
-
-
-        return NextResponse.json(
-            { message: 'User registered successfully' },
-            { status: 201 }
-        );
+        return response;
     } catch (error) {
         let message = "An error occurred";
         if (error instanceof Error) {
